requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.1.0
langchain-community==0.0.10
langchain-huggingface==0.3.1
chromadb==0.4.18
beautifulsoup4==4.12.2
jinja2==3.1.2
python-dotenv==1.0.0
streamlit==1.29.0
requests==2.31.0
pydantic==2.5.0